Move src/fmi_opendata.py -> forecasts/weather_forecast.py
